refactor (gravsearch): move transformer class for count query into separate file
