+++ webapi/src/main/scala/org/knora/webapi/responders/v2/SearchResponderV2.scala
@@ -25,7 +25,6 @@ import org.knora.webapi.messages.admin.responder.usersmessages.UserADM
 import org.knora.webapi.messages.store.triplestoremessages._
 import org.knora.webapi.messages.v2.responder.resourcemessages._
 import org.knora.webapi.messages.v2.responder.searchmessages._
-import org.knora.webapi.messages.v2.responder.valuemessages.DateValueContentV2
 import org.knora.webapi.util.ActorUtil._
 import org.knora.webapi.util.IriConversions._
 import org.knora.webapi.util._
@@ -34,8 +33,6 @@ import org.knora.webapi.util.search._
 import org.knora.webapi.util.search.gravsearch.GravsearchUtilV2.Gravsearch.GravsearchConstants
 import org.knora.webapi.util.search.gravsearch.GravsearchUtilV2.SparqlTransformation._
 import org.knora.webapi.util.search.gravsearch._
-
-import scala.collection.mutable
 import scala.concurrent.Future
 
 /**
@@ -384,78 +381,6 @@ class SearchResponderV2 extends ResponderWithStandoffV2 {
         // make sure that OFFSET is 0
         if (inputQuery.offset != 0) throw GravsearchException(s"OFFSET must be 0 for a count query, but ${inputQuery.offset} given")
 
-        /**
-          * Transforms a preprocessed CONSTRUCT query into a SELECT query that returns only the IRIs and sort order of the main resources that matched
-          * the search criteria. This query will be used to get resource IRIs for a single page of results. These IRIs will be included in a CONSTRUCT
-          * query to get the actual results for the page.
-          *
-          * @param typeInspectionResult the result of type inspection of the original query.
-          */
-        class NonTriplestoreSpecificConstructToSelectTransformer(typeInspectionResult: GravsearchTypeInspectionResult, querySchema: ApiV2Schema) extends AbstractSparqlTransformer(typeInspectionResult, querySchema) with ConstructToSelectTransformer {
-
-            def handleStatementInConstruct(statementPattern: StatementPattern): Unit = {
-                // Just identify the main resource variable and put it in mainResourceVariable.
-
-                isMainResourceVariable(statementPattern) match {
-                    case Some(queryVariable: QueryVariable) => mainResourceVariable = Some(queryVariable)
-                    case None => ()
-                }
-            }
-
-            def transformStatementInWhere(statementPattern: StatementPattern, inputOrderBy: Seq[OrderCriterion]): Seq[QueryPattern] = {
-
-                // Include any statements needed to meet the user's search criteria, but not statements that would be needed for permission checking or
-                // other information about the matching resources or values.
-
-                processStatementPatternFromWhereClause(
-                    statementPattern = statementPattern,
-                    inputOrderBy = inputOrderBy
-                )
-
-            }
-
-            def transformFilter(filterPattern: FilterPattern): Seq[QueryPattern] = {
-                val filterExpression: TransformedFilterPattern = transformFilterPattern(filterPattern.expression, typeInspectionResult = typeInspectionResult, isTopLevel = true)
-
-                filterExpression.expression match {
-                    case Some(expression: Expression) => filterExpression.additionalPatterns :+ FilterPattern(expression)
-
-                    case None => filterExpression.additionalPatterns // no FILTER expression given
-                }
-
-            }
-
-            def getSelectVariables: Seq[SelectQueryColumn] = {
-
-                val mainResVar = mainResourceVariable match {
-                    case Some(mainVar: QueryVariable) => mainVar
-
-                    case None => throw GravsearchException(s"No ${OntologyConstants.KnoraBase.IsMainResource.toSmartIri.toSparql} found in CONSTRUCT query.")
-                }
-
-                // return count aggregation function for main variable
-                Seq(Count(inputVariable = mainResVar, distinct = true, outputVariableName = "count"))
-            }
-
-            def getGroupBy(orderByCriteria: TransformedOrderBy): Seq[QueryVariable] = {
-                Seq.empty[QueryVariable]
-            }
-
-            def getOrderBy(inputOrderBy: Seq[OrderCriterion]): TransformedOrderBy = {
-                // empty by default
-                TransformedOrderBy()
-            }
-
-            def getLimit: Int = 1 // one row expected for count query
-
-            def getOffset(inputQueryOffset: Long, limit: Int): Long = {
-                // count queries do not consider offsets since there is only one result row
-                0
-            }
-
-        }
-
-
         for {
 
             // Do type inspection and remove type annotations from the WHERE clause.
@@ -471,7 +396,7 @@ class SearchResponderV2 extends ResponderWithStandoffV2 {
 
             // Create a Select prequery
 
-            nonTriplestoreSpecificConstructToSelectTransformer: NonTriplestoreSpecificConstructToSelectTransformer = new NonTriplestoreSpecificConstructToSelectTransformer(
+            nonTriplestoreSpecificConstructToSelectTransformer: NonTriplestoreSpecificConstructToSelectTransformerCountQuery = new NonTriplestoreSpecificConstructToSelectTransformerCountQuery(
                 typeInspectionResult = typeInspectionResult,
                 querySchema = inputQuery.querySchema.getOrElse(throw AssertionException(s"WhereClause has no querySchema"))
             )
